Removed mapper-attachments plugin. Now use the ingest-attachment plugin
