+++ test/Elastica/IndexTest.php
@@ -120,172 +120,6 @@ class IndexTest extends BaseTest
 
     /**
      * @group functional
-     */
-    public function testAddPdfFile()
-    {
-        $this->markTestSkipped('ES6 update: use ingest attachment : No handler for type [attachment] declared on field [file]');
-        $indexMapping = ['file' => ['type' => 'attachment'], 'text' => ['type' => 'text']];
-
-        $indexParams = ['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]];
-
-        $index = $this->_createIndex();
-        $type = new Type($index, 'test');
-
-        $index->create($indexParams, true);
-        $type->setMapping($indexMapping);
-
-        $doc1 = new Document(1);
-        $doc1->addFile('file', BASE_PATH.'/data/test.pdf', 'application/pdf');
-        $doc1->set('text', 'basel world');
-        $type->addDocument($doc1);
-
-        $doc2 = new Document(2);
-        $doc2->set('text', 'running in basel');
-        $type->addDocument($doc2);
-
-        $index->forcemerge();
-
-        $resultSet = $type->search('xodoa');
-        $this->assertEquals(1, $resultSet->count());
-
-        $resultSet = $type->search('basel');
-        $this->assertEquals(2, $resultSet->count());
-
-        // Author is ruflin
-        $resultSet = $type->search('ruflin');
-        $this->assertEquals(1, $resultSet->count());
-
-        // String does not exist in file
-        $resultSet = $type->search('guschti');
-        $this->assertEquals(0, $resultSet->count());
-    }
-
-    /**
-     * @group functional
-     */
-    public function testAddPdfFileContent()
-    {
-        $this->markTestSkipped('ES6 update: use ingest attachment : No handler for type [attachment] declared on field [file]');
-        $indexMapping = ['file' => ['type' => 'attachment'], 'text' => ['type' => 'text']];
-
-        $indexParams = ['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]];
-
-        $index = $this->_createIndex();
-        $type = new Type($index, 'test');
-
-        $index->create($indexParams, true);
-        $type->setMapping($indexMapping);
-
-        $doc1 = new Document(1);
-        $doc1->addFileContent('file', file_get_contents(BASE_PATH.'/data/test.pdf'));
-        $doc1->set('text', 'basel world');
-        $type->addDocument($doc1);
-
-        $doc2 = new Document(2);
-        $doc2->set('text', 'running in basel');
-        $type->addDocument($doc2);
-
-        $index->forcemerge();
-
-        $resultSet = $type->search('xodoa');
-        $this->assertEquals(1, $resultSet->count());
-
-        $resultSet = $type->search('basel');
-        $this->assertEquals(2, $resultSet->count());
-
-        // Author is ruflin
-        $resultSet = $type->search('ruflin');
-        $this->assertEquals(1, $resultSet->count());
-
-        // String does not exist in file
-        $resultSet = $type->search('guschti');
-        $this->assertEquals(0, $resultSet->count());
-    }
-
-    /**
-     * @group functional
-     */
-    public function testAddWordxFile()
-    {
-        $this->markTestSkipped('ES6 update: use ingest attachment : No handler for type [attachment] declared on field [file]');
-        $indexMapping = ['file' => ['type' => 'attachment'], 'text' => ['type' => 'text']];
-
-        $indexParams = ['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]];
-
-        $index = $this->_createIndex();
-        $type = new Type($index, 'content');
-
-        $index->create($indexParams, true);
-        $type->setMapping($indexMapping);
-
-        $doc1 = new Document(1);
-        $doc1->addFile('file', BASE_PATH.'/data/test.docx');
-        $doc1->set('text', 'basel world');
-        $type->addDocument($doc1);
-
-        $index->forcemerge();
-        $index->refresh();
-
-        $doc2 = new Document(2);
-        $doc2->set('text', 'running in basel');
-        $type->addDocument($doc2);
-
-        $index->forcemerge();
-        $index->refresh();
-
-        $resultSet = $type->search('basel');
-        $this->assertEquals(2, $resultSet->count());
-
-        $resultSet = $type->search('ruflin');
-        $this->assertEquals(0, $resultSet->count());
-
-        $resultSet = $type->search('Xodoa');
-        $this->assertEquals(1, $resultSet->count());
-    }
-
-    /**
-     * @group functional
-     */
-    public function testExcludeFileSource()
-    {
-        $this->markTestSkipped('ES6 update: use ingest attachment : No handler for type [attachment] declared on field [file]');
-        $indexMapping = ['file' => ['type' => 'attachment'], 'text' => ['type' => 'text', 'store' => true],
-            'title' => ['type' => 'text', 'store' => true], ];
-
-        $indexParams = ['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]];
-
-        $index = $this->_createIndex();
-        $type = new Type($index, 'content');
-
-        $mapping = Mapping::create($indexMapping);
-        $mapping->setSource(['excludes' => ['file']]);
-
-        $mapping->setType($type);
-
-        $index->create($indexParams, true);
-        $type->setMapping($mapping);
-
-        $docId = 1;
-        $text = 'Basel World';
-        $title = 'No Title';
-
-        $doc1 = new Document($docId);
-        $doc1->addFile('file', BASE_PATH.'/data/test.docx');
-        $doc1->set('text', $text);
-        $doc1->set('title', $title);
-        $type->addDocument($doc1);
-
-        // Optimization necessary, as otherwise source still in realtime get
-        $index->forcemerge();
-
-        $data = $type->getDocument($docId)->getData();
-        $this->assertEquals($data['title'], $title);
-        $this->assertEquals($data['text'], $text);
-        $this->assertFalse(isset($data['file']));
-    }
-
-    /**
-     * @group functional
      * @expectedException \Elastica\Exception\ResponseException
      */
     public function testAddRemoveAlias()
